refactor: utils폴더 이동
