+++ src/App.js
@@ -5,11 +5,7 @@ import RestaurantForm from "./components/restaurantFormSection/restaurantForm/Re
 import RestaurantListModel from "./domain/RestaurantListModel.js";
 import RestaurantNavigator from "./components/restaurantListSection/restaurantNavigator/RestaurantNavigator.js";
 import RestaurantFilterSection from "./components/restaurantListSection/restaurantFilterSection/RestaurantFilterSection.js";
-import {
-  setItem,
-  RESTAURANT_LIST_KEY,
-  getItem,
-} from "./components/utils/storage.js";
+import { setItem, RESTAURANT_LIST_KEY, getItem } from "./utils/storage.js";
 import RestaurantDetail from "./components/restaurantDetail/RestaurantDetail.js";
 
 export default class App {
